requirements.txt
# LangChain + LangGraph
langgraph>=0.2.28
langchain>=0.3.7
langchain-core>=0.3.20
langchain-community>=0.3.7
langchain-groq>=0.2.1          # Groq integration (FREE)
langchain-ollama>=0.2.1        # Ollama integration (LOCAL/FREE)

# LLM Backend (Free)
groq==0.11.0                   # Groq Python SDK

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
sse-starlette==2.1.3          # SSE streaming

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# PDF
pymupdf==1.24.13

# Media
yt-dlp==2024.11.18
faster-whisper==1.0.3

# Web Search (FREE, no API key)
duckduckgo-search==6.3.7
trafilatura==1.12.2

# Embeddings (LOCAL/FREE)
sentence-transformers==3.3.1
# For CPU-only (smaller): pip install torch --index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
faiss-cpu==1.9.0

# Caching
redis==5.2.0

# Observability
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2

# Safety/PII
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.8.2                   # Required by presidio

# Tokenization
tiktoken==0.8.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
respx==0.21.1                  # Mock httpx
networkx==3.2.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
